[v240402] - Launch Arguments Fix (Program.cs)
Fixed the program.cs so that the application doesn't overwrite the launch arguments with incomplete arguments when the application self-reboots in administrator mode. If there are already launch arguments present, the application will now assume that they are from pre-admin mode. If they are invalid/incorrect, the splash.cs now handles that event.

Co-Authored-By: Ryan Walpole <69621127+RyanWalpole@users.noreply.github.com>