[v240501] - Create New Window (AdminElevate)
Create a new window (AdminElevate) to prompt users (with Stardew Valley saved on a protected folder) to launch as administrator - and give them the option to launch without it, in the event that they don't have access to an admin account or if the issue doesn't seem to affect them. Options are good, people!

Co-Authored-By: Ryan Walpole <69621127+RyanWalpole@users.noreply.github.com>